Modify permissions test to pass (See #9192)
Once the SESSIONUUID issue was cleaned up for script
repo files, another issue appeared caused by the passing
of a call context. Namely, some of the grid infrastructure
relies on the presence of the client UUID in the context
in order to lookup servants. When an explicit call context
is passed, this value is lost. Clients can copy the current
implicit context in order to maintain this value, but we
are currently not doing that.