Replace create() test method with RawAccessRequest
The create method was the last remnant of the previous
multiplicity of ways to create files. These have now
all been consolidated into "register()". This poses a
problem for testing though. It's no longer possible to
modify the file system without also updating the database.

In order to simulate a background user (non-OMERO) modifying
files within the repository with OMERO's awareness, the
RawAccessRequest command can be used to touch or remove files.

This can only be done by an administrator.