Add UnregisteredFileException
file() and fileById() now throw an omero.UnregisteredFileException
if a file exists on the filesystem but is not registered in the DB

The `file` field of the exception returns the OriginalFile object
which would be created if register(String path, String mimetype)
were called on the same path. If the object on disk is a directory,
mimetype must be null or "Directory".