Add finish() phase to IRequest (See #11045, #11054, etc)
Now that Fileset handling is happening after containers and
images, it's necessary to have a phase hook which checks that
all filesets have been deleted along with the related images.

This amounts to a trigger pre-commit, but that's not provided
by postgresql. Instead, we collect all of the fileset IDs in
the DeleteStep instances and then check that they have been
removed afterwards. The primary drawback is that this no longer
provides a GraphConstraintERR but rather just a regular ERR.

See #10846

Note: if this requirement is ever removed, the entire finish()
phase could be rolled back. Alternatively, this could replace
the ChgrpValidation steps.