PyTest: attempt to fix import test
stderr of:
```
Image:Image:5525
5525
```

is being captured and pass to subsequent methods which fails the
expected regex (`^Image:(?P<id>\d+)$`). A possibility is that the
stdout and stderr are left over from a previous test. If this is
the case, then this commit will throw an exception. Later, we may
decide to just clean the captured streams before relying on them.

See:
 * https://ci.openmicroscopy.org/view/Failing/job/OMERO-DEV-merge-integration/911/consoleFull
 * https://ci.openmicroscopy.org/view/Failing/job/OMERO-DEV-merge-integration-python/799/testReport/OmeroPy.test.integration.clitest.test_import/TestImport/testMultipleNameModelTargets___False_/
 * https://docs.pytest.org/en/2.9.0/capture.html#accessing-captured-output-from-a-test-function